reorganize project structure to be similar to function and mobile app
