+++ shpe-app-web/app/(main)/events/components/MonthView.tsx
@@ -1,4 +1,4 @@
-import { SHPEEvent } from '@/types/Events';
+import { SHPEEvent } from '@/app/types/Events';
 import {
   startOfMonth,
   endOfMonth,
@@ -50,9 +50,8 @@ const MonthView: React.FC<MonthViewProps> = ({ eventsByDate, focusDate, setFocus
       <div
         key={`${dayState}-${index}`}
         onClick={() => setFocusDate(day)}
-        className={`relative border-t-[3px] border-l-[3px] border-[#E0E0E0] ${
-          dayState != 'curr' && 'bg-gray-100 opacity-90'
-        }`}
+        className={`relative border-t-[3px] border-l-[3px] border-[#E0E0E0] ${dayState != 'curr' && 'bg-gray-100 opacity-90'
+          }`}
       >
         {/* Day of the week text */}
         {(dayState == 'prev' || (dayState == 'curr' && index + startingDayIndex < 7)) && (
