+++ shpe-app-web/app/(main)/events/page.tsx
@@ -4,8 +4,8 @@ import { useEffect, useRef, useState } from 'react';
 import Header from '../../components/Header';
 import PendingEvent from './components/PendingEvent';
 import EventCalendar from './components/EventCalendar';
-import { getEvents, getEventLogs } from '@/helpers/firebaseUtils';
-import { SHPEEvent, SHPEEventLog } from '@/types/Events';
+import { getEvents, getEventLogs } from '@/app/helpers/firebaseUtils';
+import { SHPEEvent, SHPEEventLog } from '@/app/types/Events';
 
 const Page = () => {
   const [events, setEvents] = useState<SHPEEvent[]>([]);
