+++ shpe-app-web/app/helpers/auth.ts
@@ -1,4 +1,4 @@
-import { auth } from "@/api/firebaseConfig";
+import { auth } from "@/app/api/firebaseConfig";
 import { GoogleAuthProvider, signInWithPopup, signOut } from 'firebase/auth';
 import { AppRouterInstance } from "next/dist/shared/lib/app-router-context.shared-runtime";
