+++ shpe-app-web/app/helpers/timeUtils.ts
