+++ shpe-app-web/app/types/User.ts
