+++ shpe-app-web/package.json
@@ -11,9 +11,9 @@
   "dependencies": {
     "date-fns": "^3.6.0",
     "firebase": "^10.7.1",
-    "next": "^14.0.4",
-    "react": "^18.2.0",
-    "react-dom": "^18",
+    "next": "^14.2.3",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
     "react-icons": "^5.0.1"
   },
   "devDependencies": {
@@ -23,7 +23,7 @@
     "@types/react-dom": "^18",
     "autoprefixer": "^10.0.1",
     "eslint": "^8",
-    "eslint-config-next": "14.0.3",
+    "eslint-config-next": "^14.2.3",
     "postcss": "^8",
     "tailwindcss": "^3.3.0",
     "typescript": "^5"
